added client side spring data functionality to read, update and delete

+++ src/main/java/de/knightsoftnet/validationexample/client/services/PersonRestService.java
@@ -30,5 +30,5 @@ import javax.ws.rs.Produces;
 @Path(AppResourcePaths.PERSON)
 @Consumes("application/json")
 @Produces("application/json")
-public interface PersonRestService extends DataBaseRestServiceTemplate<Person, Integer> {
+public interface PersonRestService extends DataBaseRestServiceTemplate<Person, Long> {
 }
